+++ itachallenge-user/src/test/java/com/itachallenge/user/service/UserServiceImplTest.java
@@ -11,10 +11,15 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
+
+import com.itachallenge.userinteraction.document.favorite.FavoriteDocument;
+import com.itachallenge.userinteraction.repository.favorite.FavoriteRepository;
 import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
+import java.time.LocalDateTime;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
@@ -22,10 +27,13 @@ import java.util.UUID;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
+
 class UserServiceImplTest {
 
     @Mock
     private UserRepository userRepository;
+    @Mock
+    private FavoriteRepository favoriteRepository;
 
     @InjectMocks
     private UserServiceImpl userService;
@@ -73,23 +81,25 @@ class UserServiceImplTest {
     }
 
     @Test
-    void addChallengeToFavorites_ShouldReturnTrue_WhenFavoritesIsNull() {
-        UUID challengeId = UUID.randomUUID();
+    void addChallengeToFavorites_ShouldReturnTrue_WhenFavoriteDoesNotExist() {
         UUID userId = UUID.randomUUID();
-        UserDocument user = new UserDocument(userId, "testUser", null, null, null, 0);
+        UUID challengeId = UUID.randomUUID();
+        UserDocument user = new UserDocument(userId, "testUser", null, new HashSet<>(), null, 0);
 
         when(userRepository.findById(userId)).thenReturn(Mono.just(user));
-        when(userRepository.save(user)).thenReturn(Mono.just(user));
+        when(favoriteRepository.existsByUserIdAndChallengeId(userId, challengeId))
+                .thenReturn(Mono.just(false));
+        when(favoriteRepository.save(any(FavoriteDocument.class)))
+                .thenReturn(Mono.just(new FavoriteDocument()));
 
         StepVerifier.create(userService.addChallengeToFavorites(userId.toString(), challengeId.toString()))
                 .expectNext(true)
                 .verifyComplete();
 
-        assertNotNull(user.getFavoriteChallenges());
-        assertTrue(user.getFavoriteChallenges().contains(challengeId));
-
         verify(userRepository, times(1)).findById(userId);
-        verify(userRepository, times(1)).save(user);
+        verify(favoriteRepository, times(1)).existsByUserIdAndChallengeId(userId, challengeId);
+        verify(favoriteRepository, times(1)).save(any(FavoriteDocument.class));
+        verify(userRepository, never()).save(any());
     }
 
     @Test
@@ -113,26 +123,6 @@ class UserServiceImplTest {
     }
 
     @Test
-    void addChallengeToFavorites_ShouldReturnTrue_WhenFavoritesIsEmpty() {
-        UUID challengeId = UUID.randomUUID();
-        UUID userId = UUID.randomUUID();
-        UserDocument user = new UserDocument(userId, "testUser", null, new HashSet<>(), null, 0);
-
-        when(userRepository.findById(userId)).thenReturn(Mono.just(user));
-        when(userRepository.save(user)).thenReturn(Mono.just(user));
-
-        StepVerifier.create(userService.addChallengeToFavorites(userId.toString(), challengeId.toString()))
-                .expectNext(true)
-                .verifyComplete();
-
-        assertNotNull(user.getFavoriteChallenges());
-        assertTrue(user.getFavoriteChallenges().contains(challengeId));
-
-        verify(userRepository, times(1)).findById(userId);
-        verify(userRepository, times(1)).save(user);
-    }
-
-    @Test
     void addChallengeToBookmarks_ShouldReturnTrue_WhenBookmarksIsEmpty() {
         UUID challengeId = UUID.randomUUID();
         UUID userId = UUID.randomUUID();
@@ -147,27 +137,6 @@ class UserServiceImplTest {
 
         assertNotNull(user.getBookmarkChallenges());
         assertTrue(user.getBookmarkChallenges().contains(challengeId));
-
-        verify(userRepository, times(1)).findById(userId);
-        verify(userRepository, times(1)).save(user);
-    }
-
-    @Test
-    void addChallengeToFavorites_ShouldReturnTrue_WhenFavoritesHasValues() {
-        UUID challengeId = UUID.randomUUID();
-        UUID userId = UUID.randomUUID();
-        Set<UUID> favorites = new HashSet<>(Set.of(UUID.randomUUID(), UUID.randomUUID()));
-        UserDocument user = new UserDocument(userId, "testUser", null, favorites, null, 0);
-
-        when(userRepository.findById(userId)).thenReturn(Mono.just(user));
-        when(userRepository.save(user)).thenReturn(Mono.just(user));
-
-        StepVerifier.create(userService.addChallengeToFavorites(userId.toString(), challengeId.toString()))
-                .expectNext(true)
-                .verifyComplete();
-
-        assertNotNull(user.getFavoriteChallenges());
-        assertTrue(user.getFavoriteChallenges().contains(challengeId));
 
         verify(userRepository, times(1)).findById(userId);
         verify(userRepository, times(1)).save(user);
@@ -195,13 +164,15 @@ class UserServiceImplTest {
     }
 
     @Test
-    void addChallengeToFavorites_ShouldReturnFalse_WhenFavoritesAlreadyContainsChallenge() {
+    void addChallengeToFavorites_ShouldReturnFalse_WhenFavoriteAlreadyExists() {
         UUID challengeId = UUID.randomUUID();
         UUID userId = UUID.randomUUID();
         Set<UUID> favorites = new HashSet<>(Set.of(UUID.randomUUID(), UUID.randomUUID(), challengeId));
         UserDocument user = new UserDocument(userId, "testUser", null, favorites, null, 0);
 
         when(userRepository.findById(userId)).thenReturn(Mono.just(user));
+        when(favoriteRepository.existsByUserIdAndChallengeId(userId, challengeId))
+                .thenReturn(Mono.just(true));
 
         StepVerifier.create(userService.addChallengeToFavorites(userId.toString(), challengeId.toString()))
                 .expectNext(false)
@@ -211,7 +182,8 @@ class UserServiceImplTest {
         assertTrue(user.getFavoriteChallenges().contains(challengeId));
 
         verify(userRepository, times(1)).findById(userId);
-        verify(userRepository, times(0)).save(any());
+        verify(favoriteRepository, times(1)).existsByUserIdAndChallengeId(userId, challengeId);
+        verify(favoriteRepository, never()).save(any());
     }
 
     @Test
@@ -236,18 +208,22 @@ class UserServiceImplTest {
 
     @Test
     void addChallengeToFavorites_ShouldThrowNotFoundException_WhenUserNotFound() {
+        UUID userId = UUID.randomUUID();
+        UUID challengeId = UUID.randomUUID();
 
         when(userRepository.findById(any(UUID.class))).thenReturn(Mono.empty());
 
-        StepVerifier.create(userService.addChallengeToFavorites(UUID.randomUUID().toString(), UUID.randomUUID().toString()))
+        StepVerifier.create(userService.addChallengeToFavorites(userId.toString(), challengeId.toString()))
                 .expectErrorSatisfies(throwable -> {
                     assertInstanceOf(NotFoundException.class, throwable);
                     assertEquals("User not found", throwable.getMessage());
                 })
                 .verify();
 
-        verify(userRepository, times(1)).findById(any(UUID.class));
-        verify(userRepository, times(0)).save(any());
+        verify(userRepository, times(1)).findById(userId);
+        verify(favoriteRepository, never()).existsByUserIdAndChallengeId(any(), any());
+        verify(favoriteRepository, never()).save(any());
+        verify(userRepository, never()).save(any());
     }
 
     @Test
@@ -268,15 +244,20 @@ class UserServiceImplTest {
 
     @Test
     void addChallengeToFavorites_ShouldThrowBadRequestException_WhenUserUuidIsNull() {
-        StepVerifier.create(userService.addChallengeToFavorites(null, UUID.randomUUID().toString()))
+        String invalidUserId = null;
+        String validChallengeId = UUID.randomUUID().toString();
+
+        StepVerifier.create(userService.addChallengeToFavorites(invalidUserId, validChallengeId))
                 .expectErrorSatisfies(throwable -> {
                     assertInstanceOf(BadUUIDException.class, throwable);
                     assertEquals("Invalid ID format", throwable.getMessage());
                 })
                 .verify();
 
-        verify(userRepository, times(0)).findById(any(UUID.class));
-        verify(userRepository, times(0)).save(any());
+        verify(userRepository, never()).findById(Mockito.<UUID>any());
+        verify(favoriteRepository, never()).existsByUserIdAndChallengeId(any(), any());
+        verify(favoriteRepository, never()).save(any());
+        verify(userRepository, never()).save(any());
     }
 
     @Test
@@ -294,15 +275,20 @@ class UserServiceImplTest {
 
     @Test
     void addChallengeToFavorites_ShouldThrowBadRequestException_WhenChallengeUuidIsNull() {
-        StepVerifier.create(userService.addChallengeToFavorites(UUID.randomUUID().toString(), null))
+        String validUserId = UUID.randomUUID().toString();
+        String invalidChallengeId = null;
+
+        StepVerifier.create(userService.addChallengeToFavorites(validUserId, invalidChallengeId))
                 .expectErrorSatisfies(throwable -> {
                     assertInstanceOf(BadUUIDException.class, throwable);
                     assertEquals("Invalid ID format", throwable.getMessage());
                 })
                 .verify();
 
-        verify(userRepository, times(0)).findById(any(UUID.class));
-        verify(userRepository, times(0)).save(any());
+        verify(userRepository, never()).findById(Mockito.<UUID>any());
+        verify(favoriteRepository, never()).existsByUserIdAndChallengeId(any(), any());
+        verify(favoriteRepository, never()).save(any());
+        verify(userRepository, never()).save(any());
     }
 
     @Test
@@ -320,15 +306,20 @@ class UserServiceImplTest {
 
     @Test
     void addChallengeToFavorites_ShouldThrowBadRequestException_WhenUserUuidIsNotValid() {
-        StepVerifier.create(userService.addChallengeToFavorites("invalidUuid", UUID.randomUUID().toString()))
+        String invalidUserId = "invalidUuid";
+        String validChallengeId = UUID.randomUUID().toString();
+
+        StepVerifier.create(userService.addChallengeToFavorites(invalidUserId, validChallengeId))
                 .expectErrorSatisfies(throwable -> {
                     assertInstanceOf(BadUUIDException.class, throwable);
                     assertEquals("Invalid ID format", throwable.getMessage());
                 })
                 .verify();
 
-        verify(userRepository, times(0)).findById(any(UUID.class));
-        verify(userRepository, times(0)).save(any());
+        verify(userRepository, never()).findById(Mockito.<UUID>any());
+        verify(favoriteRepository, never()).existsByUserIdAndChallengeId(any(), any());
+        verify(favoriteRepository, never()).save(any());
+        verify(userRepository, never()).save(any());
     }
 
     @Test
@@ -345,16 +336,21 @@ class UserServiceImplTest {
     }
 
     @Test
-    void addChallengeToFavorites_ShouldThrowBadRequestException_WhenChallengeUuidIsNotValid() {
-        StepVerifier.create(userService.addChallengeToFavorites(UUID.randomUUID().toString(), "invalidUuid"))
+    void addChallengeToFavorites_ShouldThrowBadUUIDException_WhenChallengeUuidIsNotValid() {
+        String validUserId = UUID.randomUUID().toString();
+        String invalidChallengeId = "invalidUuid";
+
+        StepVerifier.create(userService.addChallengeToFavorites(validUserId, invalidChallengeId))
                 .expectErrorSatisfies(throwable -> {
                     assertInstanceOf(BadUUIDException.class, throwable);
                     assertEquals("Invalid ID format", throwable.getMessage());
                 })
                 .verify();
 
-        verify(userRepository, times(0)).findById(any(UUID.class));
-        verify(userRepository, times(0)).save(any());
+        verify(userRepository, never()).findById(Mockito.<UUID>any());
+        verify(favoriteRepository, never()).existsByUserIdAndChallengeId(any(), any());
+        verify(favoriteRepository, never()).save(any());
+        verify(userRepository, never()).save(any());
     }
 
     @Test
@@ -371,21 +367,22 @@ class UserServiceImplTest {
     }
 
     @Test
-    void deleteChallengeFromFavorites_ShouldReturnFalse_WhenFavoritesIsNull() {
-        UUID challengeId = UUID.randomUUID();
+    void deleteChallengeFromFavorites_ShouldReturnFalse_WhenFavoriteDoesNotExist() {
         UUID userId = UUID.randomUUID();
+        UUID challengeId = UUID.randomUUID();
         UserDocument user = new UserDocument(userId, "testUser", null, null, null, 0);
 
         when(userRepository.findById(userId)).thenReturn(Mono.just(user));
+        when(favoriteRepository.findByUserIdAndChallengeId(any(UUID.class), any(UUID.class)))
+                .thenReturn(Mono.empty());
 
         StepVerifier.create(userService.deleteChallengeFromFavorites(userId.toString(), challengeId.toString()))
                 .expectNext(false)
                 .verifyComplete();
 
-        assertNull(user.getFavoriteChallenges());
-
-        verify(userRepository, times(1)).findById(userId);
-        verify(userRepository, times(0)).save(user);
+        verify(userRepository).findById(userId);
+        verify(favoriteRepository).findByUserIdAndChallengeId(any(UUID.class), any(UUID.class));
+        verify(favoriteRepository, never()).delete(any());
     }
 
     @Test
@@ -407,25 +404,6 @@ class UserServiceImplTest {
     }
 
     @Test
-    void deleteChallengeFromFavorites_ShouldReturnFalse_WhenFavoritesIsEmpty() {
-        UUID challengeId = UUID.randomUUID();
-        UUID userId = UUID.randomUUID();
-        UserDocument user = new UserDocument(userId, "testUser", null, new HashSet<>(), null, 0);
-
-        when(userRepository.findById(userId)).thenReturn(Mono.just(user));
-
-        StepVerifier.create(userService.deleteChallengeFromFavorites(userId.toString(), challengeId.toString()))
-                .expectNext(false)
-                .verifyComplete();
-
-        assertNotNull(user.getFavoriteChallenges());
-        assertFalse(user.getFavoriteChallenges().contains(challengeId));
-
-        verify(userRepository, times(1)).findById(userId);
-        verify(userRepository, times(0)).save(user);
-    }
-
-    @Test
     void deleteChallengeFromBookmarks_ShouldReturnFalse_WhenBookmarksIsEmpty() {
         UUID challengeId = UUID.randomUUID();
         UUID userId = UUID.randomUUID();
@@ -439,26 +417,6 @@ class UserServiceImplTest {
 
         assertNotNull(user.getBookmarkChallenges());
         assertFalse(user.getBookmarkChallenges().contains(challengeId));
-
-        verify(userRepository, times(1)).findById(userId);
-        verify(userRepository, times(0)).save(user);
-    }
-
-    @Test
-    void deleteChallengeFromFavorites_ShouldReturnFalse_WhenChallengeNotInFavorites() {
-        UUID challengeId = UUID.randomUUID();
-        UUID userId = UUID.randomUUID();
-        Set<UUID> favorites = new HashSet<>(Set.of(UUID.randomUUID(), UUID.randomUUID()));
-        UserDocument user = new UserDocument(userId, "testUser", null, favorites, null, 0);
-
-        when(userRepository.findById(userId)).thenReturn(Mono.just(user));
-
-        StepVerifier.create(userService.deleteChallengeFromFavorites(userId.toString(), challengeId.toString()))
-                .expectNext(false)
-                .verifyComplete();
-
-        assertNotNull(user.getFavoriteChallenges());
-        assertFalse(user.getFavoriteChallenges().contains(challengeId));
 
         verify(userRepository, times(1)).findById(userId);
         verify(userRepository, times(0)).save(user);
@@ -485,24 +443,27 @@ class UserServiceImplTest {
     }
 
     @Test
-    void deleteChallengeFromFavorites_ShouldReturnTrue_WhenFavoritesAlreadyContainsChallenge() {
-        UUID challengeId = UUID.randomUUID();
+    void deleteChallengeFromFavorites_ShouldReturnTrue_WhenFavoriteAlreadyExists() {
+        // Arrange
         UUID userId = UUID.randomUUID();
-        Set<UUID> favorites = new HashSet<>(Set.of(UUID.randomUUID(), UUID.randomUUID(), challengeId));
-        UserDocument user = new UserDocument(userId, "testUser", null, favorites, null, 0);
+        UUID challengeId = UUID.randomUUID();
+        UserDocument user = new UserDocument(userId, "testUser", null, null, null, 0);
+        FavoriteDocument favorite = new FavoriteDocument(UUID.randomUUID(), userId, challengeId, LocalDateTime.now());
 
         when(userRepository.findById(userId)).thenReturn(Mono.just(user));
-        when(userRepository.save(user)).thenReturn(Mono.just(user));
+        when(favoriteRepository.findByUserIdAndChallengeId(userId, challengeId))
+                .thenReturn(Mono.just(favorite));
+        when(favoriteRepository.delete(any(FavoriteDocument.class))).thenReturn(Mono.empty());
 
+        // Act & Assert
         StepVerifier.create(userService.deleteChallengeFromFavorites(userId.toString(), challengeId.toString()))
                 .expectNext(true)
                 .verifyComplete();
 
-        assertNotNull(user.getFavoriteChallenges());
-        assertFalse(user.getFavoriteChallenges().contains(challengeId));
-
-        verify(userRepository, times(1)).findById(userId);
-        verify(userRepository, times(1)).save(any());
+        // Verify
+        verify(userRepository).findById(userId);
+        verify(favoriteRepository).findByUserIdAndChallengeId(userId, challengeId);
+        verify(favoriteRepository).delete(favorite);
     }
 
     @Test